+++ api/routes/event.js
@@ -1,5 +1,5 @@
 import express from 'express'
-import { getEventsInCalendar, addEventInCalendar, deleteEventInCalendar, getApproval, postApproval, confirmApproval, confirmCancel } from '../controllers/event.js'
+import { getEventsInCalendar, addEventInCalendar, deleteEventInCalendar, getApproval, postApproval, confirmApproval, confirmCancel} from '../controllers/event.js'
 import { verifyUser, verifyAdmin } from '../utils/verifyToken.js'
 
 const router = express.Router()