+++ api/utils/email.js
@@ -98,3 +98,61 @@ export const attendConfirmationEmail = async (approval, status) => {
         logger.error(err) 
     }
 }
+
+export const paymentRequestEmail = async (approval) => {
+    let recipient 
+    if (approval.status === 'Pending') {
+        logger.info('send payment request email')
+        recipient = approval.approverEmail
+    } else {
+        logger.info('send payment consent email')
+        recipient = approval.consenterEmail
+    } 
+    try {
+        const transport = makeTransport()
+        await transport.sendMail({
+            from: 'HR_MANAGER' + '<' + process.env.ACCOUNT_EMAIL + '>',
+            to: recipient,
+            subject: `[지출 결재] ${approval.department}팀 ${approval.name} 사용내용 : ${approval.reason} 사용일: ${approval.start}`,
+            html: `<h3>안녕하세요. ${approval.approverName}님</h3>
+                <p>${approval.name}님이 다음과 같이 결재 신청을 하였습니다.</p>
+                <p>이름 : ${approval.name}</p>
+                <p>사용일 : ${approval.start}</p>
+                <p>사용내용 : ${approval.reason}</p>
+                <p>사용금액 : ${approval.etc}</p>
+                <p>증빙 : ${approval.content}</p>
+                <a href=${process.env.DOMAIN}/api/payment/approval/${approval._id.toString()} class="button" style="background-color: #0071c2; border: none; border-radius: 8px; color: white; padding: 15px 15px; text-align: center; text-decoration: none; display: inline-block; font-size: 16px; margin: 5px; cursor: pointer;">승인</a>
+                <a href=${process.env.DOMAIN}/api/payment/cancel/${approval._id.toString()} class="button" style="background-color: #dc3545; border: none; border-radius: 8px; color: white; padding: 15px 15px; text-align: center; text-decoration: none; display: inline-block; font-size: 16px; margin: 5px; cursor: pointer;">취소</a>`    
+        })
+    } catch(err) {
+        logger.error(err)
+    }
+}
+
+// 합의자 승인 후 요청자에게 메일 송부
+export const paymentConfirmationEmail = async (approval, status) => {
+    logger.info('send payment confirmation email')  
+    let action
+    if (status==='Active') {
+        action = '승인'
+    } else if (status==='Cancel') {action = '반려'}
+    else {action = '취소'}
+    try {
+        const transport = makeTransport()
+        await transport.sendMail({
+            from: 'HR_MANAGER' + '<' + process.env.ACCOUNT_EMAIL + '>',
+            to: approval.email,
+            cc: cc,
+            subject: `[결재 ${action}] ${approval.department}팀 ${approval.name} 사용내용 : ${approval.reason} 사용일: ${approval.start}`,
+            html: `<h3>안녕하세요. ${approval.name}님</h3>
+                <p>${approval.name}님의 지출 결재 신청이 ${action} 되었습니다.</p>
+                <p>이름 : ${approval.name}</p>
+                <p>사용일 : ${approval.start}</p>
+                <p>사용내용 : ${approval.reason}</p>
+                <p>사용금액 : ${approval.etc}</p>` 
+        })
+    } catch(err) {
+        logger.error(err) 
+    }
+}
+